pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "depth-anything-3"
version = "0.1.0"
description = "Depth Anything 3"
readme = "README.md"
requires-python = ">=3.10, <=3.13"
license = { text = "Apache-2.0" }
authors = [{ name = "Your Name" }]

dependencies = [
    "pre-commit",
    "trimesh",
    "torch>=2",
    "torchvision",
    "einops",
    "huggingface_hub",
    "imageio",
    "numpy<2",
    "opencv-python",
    "open3d",
    "fastapi",
    "uvicorn",
    "requests",
    "typer",
    "pillow",
    "omegaconf",
    "evo",
    "e3nn",
    "moviepy",
    "plyfile",
    "pillow_heif",
    "safetensors",
    "uvicorn",
    "moviepy==1.0.3",
    "typer>=0.9.0",
    "pycolmap",
    "xformers; platform_system != 'Darwin'",
]

[project.optional-dependencies]
app = ["gradio>=5", "pillow>=9.0"] # requires that python3>=3.10
# Note: gsplat is installed via setup.py post-install hook (requires torch to be installed first)
all = ["depth-anything-3[app]"]


[project.scripts]
da3 = "depth_anything_3.cli:app"

[project.urls]
Homepage = "https://github.com/ByteDance-Seed/Depth-Anything-3"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.mypy]
plugins = ["jaxtyping.mypy_plugin"]

[tool.black]
line-length = 99
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  | \.git
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
known_third_party = ["bson","cruise","cv2","dataloader","diffusers","omegaconf","tensorflow","torch","torchvision","transformers","gsplat"]
known_first_party = ["common", "data", "models", "projects"]
sections = ["FUTURE","STDLIB","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]
skip_gitignore = true
line_length = 99
no_lines_before="THIRDPARTY"
